+++ elasticsearch_dsl/__init__.py
@@ -3,12 +3,12 @@ from .aggs import A
 from .function import SF
 from .search import Search, MultiSearch
 from .field import *
-from .document import DocType, MetaField
+from .document import DocType, MetaField, InnerDoc
 from .mapping import Mapping
 from .index import Index, IndexTemplate
 from .analysis import analyzer, token_filter, char_filter, tokenizer
 from .faceted_search import *
 
-VERSION = (5, 4, 0, 'dev')
+VERSION = (6, 0, 1)
 __version__ = VERSION
 __versionstr__ = '.'.join(map(str, VERSION))