+++ elasticsearch_dsl/document.py
@@ -1,25 +1,16 @@
 import collections
-import re
+from fnmatch import fnmatch
 
 from elasticsearch.exceptions import NotFoundError, RequestError
 from six import iteritems, add_metaclass
 
 from .field import Field
 from .mapping import Mapping
-from .utils import ObjectBase, AttrDict, merge
+from .utils import ObjectBase, AttrDict, merge, DOC_META_FIELDS, META_FIELDS
 from .response import HitMeta
 from .search import Search
 from .connections import connections
 from .exceptions import ValidationException, IllegalOperation
-
-DOC_META_FIELDS = frozenset((
-    'id', 'routing', 'version', 'version_type'
-))
-
-META_FIELDS = frozenset((
-    # Elasticsearch metadata fields, except 'type'
-    'index', 'using', 'score',
-)).union(DOC_META_FIELDS)
 
 
 class MetaField(object):
@@ -44,10 +35,8 @@ class DocTypeOptions(object):
         # default cluster alias, can be overriden in doc.meta
         self._using = getattr(meta, 'using', None)
 
-        # get doc_type name, if not defined take the name of the class and
-        # transform it to lower_case
-        doc_type = getattr(meta, 'doc_type',
-                re.sub(r'(.)([A-Z])', r'\1_\2', name).lower())
+        # get doc_type name, if not defined use 'doc'
+        doc_type = getattr(meta, 'doc_type', 'doc')
 
         # create the mapping instance
         self.mapping = getattr(meta, 'mapping', Mapping(doc_type))
@@ -72,6 +61,9 @@ class DocTypeOptions(object):
                 self._using = self._using or b._doc_type._using
                 self.index = self.index or b._doc_type.index
 
+        # custom method to determine if a hit belongs to this DocType
+        self._matches = getattr(meta, 'matches', None)
+
     @property
     def using(self):
         return self._using or 'default'
@@ -94,6 +86,20 @@ class DocTypeOptions(object):
 
     def refresh(self, index=None, using=None):
         self.mapping.update_from_es(index or self.index, using=using or self.using)
+
+    def matches(self, hit):
+        if self._matches is not None:
+            return self._matches(hit)
+
+        return (
+                self.index is None or fnmatch(hit.get('_index', ''), self.index)
+            ) and self.name == hit.get('_type')
+
+@add_metaclass(DocTypeMeta)
+class InnerDoc(ObjectBase):
+    """
+    Common class for inner documents like Object or Nested
+    """
 
 
 @add_metaclass(DocTypeMeta)
@@ -249,27 +255,6 @@ class DocType(ObjectBase):
             message = 'Documents %s not found.' % ', '.join(missing_ids)
             raise NotFoundError(404, message, {'docs': missing_docs})
         return objs
-
-    @classmethod
-    def from_es(cls, hit):
-        """
-        Helper method to construct an instance from a dictionary returned by
-        elasticsearch.
-        """
-        # don't modify in place
-        meta = hit.copy()
-        doc = meta.pop('_source', {})
-
-        if 'fields' in meta:
-            for k, v in iteritems(meta.pop('fields')):
-                if k == '_source':
-                    doc.update(v)
-                if k.startswith('_') and k[1:] in META_FIELDS:
-                    meta[k] = v
-                else:
-                    doc[k] = v
-
-        return cls(meta=meta, **doc)
 
     def _get_connection(self, using=None):
         return connections.get_connection(using or self._doc_type.using)