+++ elasticsearch_dsl/utils.py
@@ -10,6 +10,15 @@ from .exceptions import UnknownDslObject, ValidationException
 SKIP_VALUES = ('', None)
 EXPAND__TO_DOT=True
 
+DOC_META_FIELDS = frozenset((
+    'id', 'routing', 'version', 'version_type'
+))
+
+META_FIELDS = frozenset((
+    # Elasticsearch metadata fields, except 'type'
+    'index', 'using', 'score',
+)).union(DOC_META_FIELDS)
+
 def _wrap(val, obj_wrapper=None):
     if isinstance(val, collections.Mapping):
         return AttrDict(val) if obj_wrapper is None else obj_wrapper(val)
@@ -326,11 +335,24 @@ class DslBase(object):
 
 class ObjectBase(AttrDict):
     def __init__(self, **kwargs):
-        m = self._doc_type.mapping
-        for k in m:
-            if k in kwargs and m[k]._coerce:
-                kwargs[k] = m[k].deserialize(kwargs[k])
         super(ObjectBase, self).__init__(kwargs)
+
+    @classmethod
+    def from_es(cls, hit):
+        meta = hit.copy()
+        doc = meta.pop('_source', {})
+        if 'fields' in meta:
+            for k, v in iteritems(meta.pop('fields')):
+                if k.startswith('_') and k[1:] in META_FIELDS:
+                    meta[k] = v
+                else:
+                    doc[k] = v
+
+        m = cls._doc_type.mapping
+        for k in m:
+            if k in doc and m[k]._coerce:
+                doc[k] = m[k].deserialize(doc[k])
+        return cls(meta=meta, **doc)
 
     def __getattr__(self, name):
         try:
@@ -345,11 +367,6 @@ class ObjectBase(AttrDict):
                         value = getattr(self, name)
                     return value
             raise
-
-    def __setattr__(self, name, value):
-        if name in self._doc_type.mapping:
-            value = self._doc_type.mapping[name].deserialize(value)
-        super(ObjectBase, self).__setattr__(name, value)
 
     def to_dict(self):
         out = {}