+++ test_elasticsearch_dsl/test_field.py
@@ -1,3 +1,5 @@
+import base64
+import ipaddress
 from datetime import datetime
 from dateutil import tz
 
@@ -64,20 +66,8 @@ def test_multi_fields_are_accepted_and_parsed():
         }
     } == f.to_dict()
 
-def test_modifying_nested():
-    f = field.Nested()
-    f.field('name', 'text', index='not_analyzed')
-
-    assert {
-        'type': 'nested',
-        'properties': {
-            'name': {'type': 'text', 'index': 'not_analyzed'}
-        },
-    } == f.to_dict()
-
 def test_nested_provides_direct_access_to_its_fields():
-    f = field.Nested()
-    f.field('name', 'text', index='not_analyzed')
+    f = field.Nested(properties={'name': {'type': 'text', 'index': 'not_analyzed'}})
 
     assert 'name' in f
     assert f['name'] == field.Text(index='not_analyzed')
@@ -111,3 +101,36 @@ def test_scaled_float():
         field.ScaledFloat()
     f = field.ScaledFloat(123)
     assert f.to_dict() == {'scaling_factor': 123, 'type': 'scaled_float'}
+
+
+def test_ipaddress():
+    f = field.Ip()
+    assert f.deserialize('127.0.0.1') == ipaddress.ip_address(u'127.0.0.1')
+    assert f.deserialize(u'::1') == ipaddress.ip_address(u'::1')
+    assert f.serialize(f.deserialize('::1')) == '::1'
+    assert f.deserialize(None) is None
+    with pytest.raises(ValueError):
+        assert f.deserialize('not_an_ipaddress')
+
+
+def test_float():
+    f = field.Float()
+    assert f.deserialize('42') == 42.0
+    assert f.deserialize(None) is None
+    with pytest.raises(ValueError):
+        assert f.deserialize('not_a_float')
+
+
+def test_integer():
+    f = field.Integer()
+    assert f.deserialize('42') == 42
+    assert f.deserialize(None) is None
+    with pytest.raises(ValueError):
+        assert f.deserialize('not_an_integer')
+
+
+def test_binary():
+    f = field.Binary()
+    assert f.deserialize(base64.b64encode(b'42')) == b'42'
+    assert f.deserialize(f.serialize(b'42')) == b'42'
+    assert f.deserialize(None) is None